Uploaded Program1.
Created a Java program with the main function in file Program1.java. This program accepts data from the user (Scanner) to create 3 objects of type Album, each with all 4 values (attributes) as required in the class. You may pass the 4 attributes to the new object or you may use the default to create the default object then set each of the 4 attributes separately.

Met specs listed in the assignment document.